requirements.txt
# GeoSense Platform - Python Dependencies
# For pip installation: pip install -r requirements.txt
# Prefer using: pip install -e ".[dev,ml]" with pyproject.toml

# Core numerical computing
jax[cpu]==0.4.20
jaxlib==0.4.20
numpy>=1.24.0,<2.0.0
scipy==1.11.0

# Data handling
pandas==2.0.0
xarray==2023.1.0
h5py==3.9.0
netCDF4==1.6.4

# Visualization
matplotlib==3.7.0
seaborn==0.12.0
cartopy==0.22.0

# Configuration
pyyaml==6.0
pydantic==2.0.0
python-dotenv==1.0.0

# Web framework
fastapi==0.100.0
uvicorn[standard]==0.23.0
httpx==0.24.0
slowapi==0.1.9  # Rate limiting
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-multipart==0.0.6  # File uploads

# Task queue
celery==5.3.0
redis==5.0.0

# Database
sqlalchemy==2.0.0
psycopg2-binary==2.9.6
asyncpg==0.28.0
alembic==1.12.0  # Database migrations

# Object Storage
minio==7.1.17

# Development tools
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
mypy==1.5.0
ruff==0.1.0
black==23.7.0
isort==5.12.0
pre-commit==3.3.0

# Jupyter
ipython==8.14.0
jupyter==1.0.0
jupyterlab==4.0.0

# Machine Learning (optional)
torch==2.0.0
flax==0.7.0
optax==0.1.7
wandb==0.15.0
tensorboard==2.14.0

# Monitoring
prometheus-client==0.17.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
